fix: bring mhgu scaffold up on CachyOS + Go 1.26
Three classes of compile error prevented 'go build' from succeeding.

1. log/slog has no Uint32 helper. The stdlib package only exports
   Uint64 (and Any). Five call sites in main.go and datastore.go
   used slog.Uint32("...", x) for c.ID, req.Method, and req.CallID;
   rewrote each as slog.Uint64("...", uint64(x)). Pre-existing
   scaffold bug, not a regression.

2. mhgu_utility.go carried an unused local 's := conn.Settings'
   declared inside the wrapper closure, left over from a planned
   override pattern that was never added. Removed; re-add the line
   the day MHGU-specific Utility overrides land.

3. go.mod now points at the local nextendo-nex clone via a
   'replace' directive (../../work/nextendo/nextendo-nex) so
   'go mod tidy' resolves without network access during bring-up.

The local nextendo-nex clone also carries two fixes:
- natbridge.go: Go 1.26 rejected adjacent string literals joined
  by '+' across a newline; collapsed into one literal.
- reaper.go: unused 'sync' import (the only sync type is the
  reaperOnce field on Endpoint, declared in endpoint.go:81).

Neither upstream fix lives in this repo and both will be lost if the
local clone is refreshed. They are noted here so the next person
who runs 'go mod tidy' does not get stuck on them again.

'go vet ./...' and 'go build -o mhgu-server .' both exit 0 on Go 1.26.5.

+++ datastore.go
@@ -51,7 +51,7 @@ func datastoreStubHandler() nex.RMCHandler {
 		s := conn.Settings
 		logger.Debug("datastore stub",
 			slog.Uint64("pid", conn.PID),
-			slog.Uint32("method", req.Method),
+			slog.Uint64("method", uint64(req.Method)),
 			slog.Int("bodyLen", len(req.Body)))
 		return nex.NewRMCSuccess(s, 0x73, req.Method, req.CallID, nil)
 	}
